+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+### Changed
+- [INTERNAL] Prepare update to vscode 1.65.2
+
+
 ## [1.3.14] - 20220316
 ### Changed
 - [INTERNAL] Prepare update to vscode 1.60.2